directly call regression method objects from Python
move the old entry points to test area
